refactor: add  draft and related classes to add Handling Event use-case
